Separated Small versions of LangSelect & Menu to own files
