+++ lib/LangSelect/LangSelect.js
@@ -3,14 +3,9 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.LangSelectSmall = undefined;
 
-var _templateObject = _taggedTemplateLiteral(['\n  border-radius: 3px;\n  border: 0;\n  outline: 0;\n  background: 0;\n  padding: 0 0.5rem;\n  color: inherit;\n  font: inherit;\n  &:hover {\n    cursor: pointer;\n  }\n  ', '\n'], ['\n  border-radius: 3px;\n  border: 0;\n  outline: 0;\n  background: 0;\n  padding: 0 0.5rem;\n  color: inherit;\n  font: inherit;\n  &:hover {\n    cursor: pointer;\n  }\n  ', '\n']),
-    _templateObject2 = _taggedTemplateLiteral(['\n  display: flex;\n  justify-content: space-between;\n  height: 100%;\n'], ['\n  display: flex;\n  justify-content: space-between;\n  height: 100%;\n']),
-    _templateObject3 = _taggedTemplateLiteral([''], ['']),
-    _templateObject4 = _taggedTemplateLiteral(['\n  ', ' {\n    display: block;\n  }\n'], ['\n  ', ' {\n    display: block;\n  }\n']),
-    _templateObject5 = _taggedTemplateLiteral(['\n  ', '\n  position: absolute;\n  border-radius: 3px;\n  ', '\n  ', ' {\n    &:hover {\n\n    }\n  }\n'], ['\n  ', '\n  position: absolute;\n  border-radius: 3px;\n  ', '\n  ', ' {\n    &:hover {\n\n    }\n  }\n']),
-    _templateObject6 = _taggedTemplateLiteral(['\n  padding: 0;\n  ', '\n  .selected-language {\n    svg {\n      margin-left: 0.25rem;\n    }\n  }\n'], ['\n  padding: 0;\n  ', '\n  .selected-language {\n    svg {\n      margin-left: 0.25rem;\n    }\n  }\n']);
+var _templateObject = _taggedTemplateLiteral(['\n  display: flex;\n  justify-content: space-between;\n  height: 100%;\n'], ['\n  display: flex;\n  justify-content: space-between;\n  height: 100%;\n']),
+    _templateObject2 = _taggedTemplateLiteral([''], ['']);
 
 var _react = require('react');
 
@@ -24,17 +19,15 @@ var _styledComponents = require('styled-components');
 
 var _styledComponents2 = _interopRequireDefault(_styledComponents);
 
-var _Icons = require('../Icons');
+var _LangButton = require('./LangButton');
+
+var _LangButton2 = _interopRequireDefault(_LangButton);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _taggedTemplateLiteral(strings, raw) { return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }
 
-var LangButton = _styledComponents2.default.button(_templateObject, function (props) {
-  return props.active && props.theme.radioBtnActive && 'background-color: ' + props.theme.radioBtnActive + ';';
-});
-
-var FlexWrapper = _styledComponents2.default.div(_templateObject2);
+var FlexWrapper = _styledComponents2.default.div(_templateObject);
 
 var LangSelect = function LangSelect(_ref) {
   var languages = _ref.languages,
@@ -46,7 +39,7 @@ var LangSelect = function LangSelect(_ref) {
     { className: className },
     languages.map(function (lang) {
       return _react2.default.createElement(
-        LangButton,
+        _LangButton2.default,
         {
           key: lang.id,
           active: selectedLanguage === lang.id,
@@ -70,69 +63,4 @@ LangSelect.propTypes = {
   className: _propTypes2.default.string
 };
 
-exports.default = (0, _styledComponents2.default)(LangSelect)(_templateObject3);
-
-
-var Div = _styledComponents2.default.div(_templateObject4, LangButton);
-
-var SelectWrapper = _styledComponents2.default.div(_templateObject5, function (props) {
-  return props.theme.background && 'background: ' + props.theme.background + ';';
-}, function (props) {
-  return props.theme.menuBorder && 'border: 1px solid ' + props.theme.menuBorder + ';';
-}, LangButton);
-
-var LangSelectSmallUnstyled = function LangSelectSmallUnstyled(_ref2) {
-  var languages = _ref2.languages,
-      changeLanguage = _ref2.changeLanguage,
-      selectedLanguage = _ref2.selectedLanguage,
-      fill = _ref2.fill,
-      open = _ref2.open,
-      className = _ref2.className;
-  return _react2.default.createElement(
-    Div,
-    { className: className },
-    _react2.default.createElement(
-      LangButton,
-      { className: 'selected-language' },
-      languages.find(function (lang) {
-        return lang.id === selectedLanguage;
-      }).name,
-      _react2.default.createElement(_Icons.ArrowDown, { height: '0.75rem', width: '0.75rem', fill: fill })
-    ),
-    open && _react2.default.createElement(
-      SelectWrapper,
-      { className: 'select-wrapper' },
-      languages.filter(function (lang) {
-        return lang.id !== selectedLanguage;
-      }).map(function (lang) {
-        return _react2.default.createElement(
-          LangButton,
-          {
-            key: lang.id,
-            onClick: function onClick() {
-              changeLanguage(lang.id);
-            }
-          },
-          lang.name
-        );
-      })
-    )
-  );
-};
-
-LangSelectSmallUnstyled.propTypes = {
-  languages: _propTypes2.default.arrayOf(_propTypes2.default.shape({
-    id: _propTypes2.default.oneOfType([_propTypes2.default.string, _propTypes2.default.number]).isRequired,
-    name: _propTypes2.default.string.isRequired
-  })).isRequired,
-  changeLanguage: _propTypes2.default.func.isRequired,
-  selectedLanguage: _propTypes2.default.oneOfType([_propTypes2.default.string, _propTypes2.default.number]),
-  fill: _propTypes2.default.string.isRequired,
-  open: _propTypes2.default.bool,
-  className: _propTypes2.default.string
-};
-var LangSelectSmall = (0, _styledComponents2.default)(LangSelectSmallUnstyled)(_templateObject6, function (props) {
-  return props.theme.background && 'background: ' + props.theme.background + ';';
-});
-
-exports.LangSelectSmall = LangSelectSmall;
+exports.default = (0, _styledComponents2.default)(LangSelect)(_templateObject2);
