+++ lib/LangSelect/index.js
@@ -1,0 +1,19 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.LangSelectSmall = undefined;
+
+var _LangSelect = require('./LangSelect');
+
+var _LangSelect2 = _interopRequireDefault(_LangSelect);
+
+var _LangSelectSmall = require('./LangSelectSmall');
+
+var _LangSelectSmall2 = _interopRequireDefault(_LangSelectSmall);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+exports.default = _LangSelect2.default;
+exports.LangSelectSmall = _LangSelectSmall2.default;
