+++ lib/Menu/MenuSmall.js
@@ -4,8 +4,7 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
-var _templateObject = _taggedTemplateLiteral([''], ['']),
-    _templateObject2 = _taggedTemplateLiteral(['\n  .lang-select {\n    height: 1.75rem;\n    width: 7rem;\n    .small {\n      display: none;\n    }\n  }\n'], ['\n  .lang-select {\n    height: 1.75rem;\n    width: 7rem;\n    .small {\n      display: none;\n    }\n  }\n']);
+var _templateObject = _taggedTemplateLiteral(['\n  > * {\n    ', '\n    padding: 0.5rem;\n  }\n'], ['\n  > * {\n    ', '\n    padding: 0.5rem;\n  }\n']);
 
 var _react = require('react');
 
@@ -21,15 +20,7 @@ var _styledComponents2 = _interopRequireDefault(_styledComponents);
 
 var _LangSelect = require('../LangSelect');
 
-var _LangSelect2 = _interopRequireDefault(_LangSelect);
-
 var _Icons = require('../Icons');
-
-var _Separator = require('../Separator/Separator');
-
-var _Span = require('../Span/Span');
-
-var _Span2 = _interopRequireDefault(_Span);
 
 var _FlexWrapper = require('../FlexWrapper/FlexWrapper');
 
@@ -39,9 +30,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 function _taggedTemplateLiteral(strings, raw) { return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }
 
-var Div = _styledComponents2.default.div(_templateObject);
-
-var Menu = function Menu(_ref) {
+var MenuSmall = function MenuSmall(_ref) {
   var languages = _ref.languages,
       changeLanguage = _ref.changeLanguage,
       selectedLanguage = _ref.selectedLanguage,
@@ -51,41 +40,26 @@ var Menu = function Menu(_ref) {
       className = _ref.className,
       children = _ref.children;
   return _react2.default.createElement(
-    Div,
+    _FlexWrapper2.default,
     { className: className },
-    _react2.default.createElement(
-      _FlexWrapper2.default,
-      { className: 'top' },
-      _react2.default.createElement(_Icons.Search, { fill: iconFill, height: searchIcon.height, width: searchIcon.width }),
-      _react2.default.createElement(
-        _Span2.default,
-        { className: 'lang-select' },
-        _react2.default.createElement(_LangSelect2.default, {
-          className: 'large',
-          languages: languages,
-          selectedLanguage: selectedLanguage,
-          changeLanguage: changeLanguage
-        }),
-        _react2.default.createElement(_LangSelect.LangSelectSmall, {
-          className: 'small',
-          languages: languages,
-          selectedLanguage: selectedLanguage,
-          changeLanguage: changeLanguage,
-          open: langSelectOpen,
-          fill: iconFill
-        })
-      )
-    ),
-    _react2.default.createElement(_Separator.MenuSeparator, null),
-    _react2.default.createElement(
-      _FlexWrapper2.default,
-      { className: 'bottom' },
-      children
-    )
+    children,
+    _react2.default.createElement(_Icons.Search, {
+      className: 'search-icon',
+      fill: iconFill,
+      height: searchIcon.height,
+      width: searchIcon.width
+    }),
+    _react2.default.createElement(_LangSelect.LangSelectSmall, {
+      languages: languages,
+      selectedLanguage: selectedLanguage,
+      changeLanguage: changeLanguage,
+      open: langSelectOpen,
+      fill: iconFill
+    })
   );
 };
 
-Menu.propTypes = {
+MenuSmall.propTypes = {
   languages: _propTypes2.default.arrayOf(_propTypes2.default.shape({
     id: _propTypes2.default.oneOfType([_propTypes2.default.string, _propTypes2.default.number]).isRequired,
     name: _propTypes2.default.string.isRequired
@@ -102,4 +76,6 @@ Menu.propTypes = {
   children: _propTypes2.default.node
 };
 
-exports.default = (0, _styledComponents2.default)(Menu)(_templateObject2);
+exports.default = (0, _styledComponents2.default)(MenuSmall)(_templateObject, function (props) {
+  return props.theme.menuBorder && 'border-left: 1px solid ' + props.theme.menuBorder + ';';
+});
