+++ lib/Menu/index.js
@@ -1,0 +1,19 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.MenuSmall = undefined;
+
+var _Menu = require('./Menu');
+
+var _Menu2 = _interopRequireDefault(_Menu);
+
+var _MenuSmall = require('./MenuSmall');
+
+var _MenuSmall2 = _interopRequireDefault(_MenuSmall);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+exports.default = _Menu2.default;
+exports.MenuSmall = _MenuSmall2.default;
