+++ lib/index.js
@@ -45,11 +45,11 @@ var _Span = require('./Span/Span');
 
 var _Span2 = _interopRequireDefault(_Span);
 
-var _LangSelect = require('./LangSelect/LangSelect');
+var _LangSelect = require('./LangSelect');
 
 var _LangSelect2 = _interopRequireDefault(_LangSelect);
 
-var _Menu = require('./Menu/Menu');
+var _Menu = require('./Menu');
 
 var _Menu2 = _interopRequireDefault(_Menu);
