+++ src/FlexWrapper/FlexWrapper.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const FlexWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+export default FlexWrapper;
