+++ src/LangSelect/LangSelect.js
@@ -2,21 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-import { ArrowDown } from '../Icons';
-
-const LangButton = styled.button`
-  border-radius: 3px;
-  border: 0;
-  outline: 0;
-  background: 0;
-  padding: 0 0.5rem;
-  color: inherit;
-  font: inherit;
-  &:hover {
-    cursor: pointer;
-  }
-  ${props => (props.active && props.theme.radioBtnActive && `background-color: ${props.theme.radioBtnActive};`)}
-`;
+import LangButton from './LangButton';
 
 const FlexWrapper = styled.div`
   display: flex;
@@ -66,82 +52,3 @@ LangSelect.propTypes = {
 
 export default styled(LangSelect)``;
 
-const Div = styled.div`
-  ${LangButton} {
-    display: block;
-  }
-`;
-
-const SelectWrapper = styled.div`
-  ${props => (props.theme.background && `background: ${props.theme.background};`)}
-  position: absolute;
-  border-radius: 3px;
-  ${props => (props.theme.menuBorder && `border: 1px solid ${props.theme.menuBorder};`)}
-  ${LangButton} {
-    &:hover {
-
-    }
-  }
-`;
-
-const LangSelectSmallUnstyled = ({
-  languages,
-  changeLanguage,
-  selectedLanguage,
-  fill,
-  open,
-  className
-}) => (
-  <Div className={className}>
-    <LangButton className="selected-language">
-      { languages.find(lang => lang.id === selectedLanguage).name }
-      <ArrowDown height="0.75rem" width="0.75rem" fill={fill} />
-    </LangButton>
-    {
-     open && <SelectWrapper className="select-wrapper">
-       {
-        languages.filter(lang => lang.id !== selectedLanguage).map(lang =>
-          (
-            <LangButton
-              key={lang.id}
-              onClick={() => { changeLanguage(lang.id); }}
-            >
-              {lang.name}
-            </LangButton>
-          )
-        )
-      }
-     </SelectWrapper>
-    }
-  </Div>
-);
-
-
-LangSelectSmallUnstyled.propTypes = {
-  languages: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.oneOfType([
-      PropTypes.string,
-      PropTypes.number,
-    ]).isRequired,
-    name: PropTypes.string.isRequired
-  })).isRequired,
-  changeLanguage: PropTypes.func.isRequired,
-  selectedLanguage: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-  ]),
-  fill: PropTypes.string.isRequired,
-  open: PropTypes.bool,
-  className: PropTypes.string
-};
-const LangSelectSmall = styled(LangSelectSmallUnstyled)`
-  padding: 0;
-  ${props => (props.theme.background && `background: ${props.theme.background};`)}
-  .selected-language {
-    svg {
-      margin-left: 0.25rem;
-    }
-  }
-`;
-
-export { LangSelectSmall };
