+++ src/LangSelect/index.js
@@ -1,0 +1,7 @@
+import LangSelect from './LangSelect';
+import LangSelectSmall from './LangSelectSmall';
+
+export default LangSelect;
+export {
+  LangSelectSmall
+};
