+++ src/Menu/index.js
@@ -1,0 +1,7 @@
+import Menu from './Menu';
+import MenuSmall from './MenuSmall';
+
+export default Menu;
+export {
+  MenuSmall
+};
